split urwid-using bits out of subiquitycore.core.Application
