+++ console_conf/core.py
@@ -15,7 +15,7 @@
 
 import logging
 
-from subiquitycore.core import Application
+from subiquitycore.tui import TuiApplication
 
 from console_conf.models.console_conf import ConsoleConfModel
 from console_conf.models.systems import RecoverySystemsModel
@@ -23,7 +23,7 @@ from console_conf.models.systems import RecoverySystemsModel
 log = logging.getLogger("console_conf.core")
 
 
-class ConsoleConf(Application):
+class ConsoleConf(TuiApplication):
 
     project = "console_conf"
 
@@ -36,7 +36,7 @@ class ConsoleConf(Application):
     ]
 
 
-class RecoveryChooser(Application):
+class RecoveryChooser(TuiApplication):
 
     project = "console_conf"
